requirements.txt
discord.py==2.6.4
aiohttp>=3.9.0
python-dotenv>=1.0.1
